+++ CHANGELOG.md
@@ -83,3 +83,20 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Security
 - Nothing
+
+## [1.2.2] - 2020-01-30
+
+### Added
+- Insert query
+
+### Deprecated
+- Nothing
+
+### Fixed
+- Examples. Warning and ambiguity.
+
+### Removed
+- Nothing
+
+### Security
+- Nothing